Applied patch from Terry Smith for JETTY-126 handle content > Integer.MAX_VALUE

+++ jetty/trunk/modules/jetty/src/test/java/org/mortbay/jetty/HttpGeneratorTest.java
@@ -264,7 +264,7 @@ public class HttpGeneratorTest extends TestCase
             content= null;
         }
 
-        public void messageComplete(int contentLength)
+        public void messageComplete(long contentLength)
         {
         }
